refactor: 식습관 분석 로직에 존재하는 PromptTemplate 내용 templates 디렉토리로 이전
